data_pipeline/pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "lxml",
    "unstructured[html]",
    "python-dotenv",
    "supabase",
]
